requirements.txt
requests~=2.32.3
PyYAML~=5.3.1
urllib3~=2.2.3
PyMySQL~=0.9.3
pytest~=8.3.4